disabling

+++ event_start.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
+
 source config.sh
-rm -f event-pics
-touch event-pics
-./send_notification.sh "EVENT START" 60s
+
+if [ ! -f disabled ]; then
+    rm -f event-pics
+    touch event-pics
+    ./send_notification.sh "EVENT START" 60s
+fi
